# internal/core/transport/http/response/handler.go
package core_http_response

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	core_errors "github.com/StanislavSizhuk/go_to_do/internal/core/errors"
	core_logger "github.com/StanislavSizhuk/go_to_do/internal/core/logger"
	"go.uber.org/zap"
)

// ErrorResponse is the JSON body returned for failed requests.
type ErrorResponse struct {
	Message string `json:"message" example:"failed to get user"`
	Error   string `json:"error" example:"not found"`
}

type HTTPREsponseHandler struct {
	log *core_logger.Logger
	rw  http.ResponseWriter
}

func NewHTTPResponseHandler(log *core_logger.Logger, rw http.ResponseWriter) *HTTPREsponseHandler {
	return &HTTPREsponseHandler{
		log: log,
		rw:  rw,
	}
}

func (h *HTTPREsponseHandler) JSONResponse(responseBody any, statusCode int) {
	h.rw.WriteHeader(statusCode)
	if err := json.NewEncoder(h.rw).Encode(responseBody); err != nil {
		h.log.Error("write HTTP response", zap.Error(err))
	}
}

func (h *HTTPREsponseHandler) NoContentResponse() {
	h.rw.WriteHeader(http.StatusNoContent)
}

func (h *HTTPREsponseHandler) ErrorResponse (err error, msg string) {

	var(
		statusCode int
		logFunc func(string, ...zap.Field)
	)

	switch{
	case errors.Is(err, core_errors.ErrInvalidArgument):
		statusCode = http.StatusBadRequest
		logFunc = h.log.Warn

	case errors.Is(err, core_errors.ErrNotFound):
		statusCode = http.StatusNotFound

		logFunc  = h.log.Debug

	
	case errors.Is(err, core_errors.ErrConflict):
		statusCode = http.StatusConflict
		logFunc = h.log.Warn
	
	default:
		statusCode = http.StatusInternalServerError
		logFunc = h.log.Error
	}

	logFunc(msg, zap.Error(err))

	h.errorResponse(
		statusCode,
		err,
		msg, 
	)
	//h.errorResponse( , err, msg)

}



func (h *HTTPREsponseHandler) PanicResponse(p any, msg string) {
	statusCode := http.StatusInternalServerError
	err := fmt.Errorf("unexpected panic: %v", p)

	h.log.Error(msg, zap.Error(err))

	h.errorResponse(
		statusCode,
		err,
		msg,
	)
}

func (h *HTTPREsponseHandler) errorResponse(
	statusCode int,
	err error,
	msg string,
){

	response := map[string]string{
		"message": msg,
		"error":   err.Error(),
	}

	h.JSONResponse(
		response,
		statusCode, 
	)

	
}